separate node and browser tests - event-hubs version
