+++ sdk/eventhub/event-hubs/test/client.spec.ts
@@ -14,7 +14,6 @@ import { TokenCredential, earliestEventPosition } from "../src";
 import { EventHubClient } from "../src/impl/eventHubClient";
 import { packageJsonInfo } from "../src/util/constants";
 import { EnvVarKeys, getEnvVars, isNode } from "./utils/testUtils";
-import { EnvironmentCredential } from "@azure/identity";
 import { EventHubConsumer } from "../src/receiver";
 import { EventHubProducer } from "../src/sender";
 const env = getEnvVars();
@@ -82,47 +81,6 @@ describe("Create EventHubClient using connection string", function(): void {
     const client = new EventHubClient("abc", "my-event-hub-name", dummyCredential);
     client.should.be.an.instanceof(EventHubClient);
     should.equal(client.eventHubName, "my-event-hub-name");
-  });
-});
-
-describe("Create EventHubClient using Azure Identity", function(): void {
-  it("creates an EventHubClient from an Azure.Identity credential", async function(): Promise<
-    void
-  > {
-    should.exist(
-      env[EnvVarKeys.AZURE_CLIENT_ID],
-      "define AZURE_CLIENT_ID in your environment before running integration tests."
-    );
-    should.exist(
-      env[EnvVarKeys.AZURE_TENANT_ID],
-      "define AZURE_TENANT_ID in your environment before running integration tests."
-    );
-    should.exist(
-      env[EnvVarKeys.AZURE_CLIENT_SECRET],
-      "define AZURE_CLIENT_SECRET in your environment before running integration tests."
-    );
-    should.exist(
-      env[EnvVarKeys.EVENTHUB_CONNECTION_STRING],
-      "define EVENTHUB_CONNECTION_STRING in your environment before running integration tests."
-    );
-
-    // This is of the form <your-namespace>.servicebus.windows.net
-    const endpoint = (env.EVENTHUB_CONNECTION_STRING.match("Endpoint=sb://(.*)/;") || "")[1];
-
-    const credential = new EnvironmentCredential();
-    const client = new EventHubClient(endpoint, env.EVENTHUB_NAME, credential);
-
-    // Extra check involving actual call to the service to ensure this works
-    const hubInfo = await client.getProperties();
-    should.equal(hubInfo.name, client.eventHubName);
-    await client.close();
-  });
-
-  it("Verify fullyQualifiedNamespace when creating an EventHubClient from an Azure.Identity credential", function(): void {
-    const endpoint = "test.servicebus.windows.net";
-    const credential = new EnvironmentCredential();
-    const client = new EventHubClient(endpoint, "my-event-hub-name", credential);
-    should.equal(client.fullyQualifiedNamespace, "test.servicebus.windows.net");
   });
 });
