+++ sdk/eventhub/event-hubs/test/node/packageInfo.spec.ts
@@ -5,7 +5,7 @@ import chai from "chai";
 const should = chai.should();
 import fs from "fs";
 import path from "path";
-import { packageJsonInfo } from "../src/util/constants";
+import { packageJsonInfo } from "../../src/util/constants";
 
 // Since we currently hardcode package name and version in `constants.ts` file,
 // following test is in place to ensure the values in package.json and in this file are consistent
